+++ src/NumArray/FactoryInterface.php
@@ -1,12 +1,13 @@
 <?php
+declare(strict_types=1);
 
 namespace NumPHP\NumArray;
 
 interface FactoryInterface
 {
-    public function createFromData(array $data): NumArray;
+    public function createFromData(array $data): NumArrayInterface;
 
-    public function createZeros(int ...$axis): NumArray;
+    public function createZeros(int ...$axis): NumArrayInterface;
 
-    public function createZerosLike(NumArray $numArray): NumArray;
+    public function createZerosLike(NumArrayInterface $numArray): NumArrayInterface;
 }