+++ src/NumArray/NumArrayInterface.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace NumPHP\NumArray;
+
+interface NumArrayInterface
+{
+    public function __toString(): string;
+
+    public function getData(): array;
+
+    public function getShape(): array;
+
+    public function getSize(): int;
+
+    public function getNDim(): int;
+}